requirements.txt
jinja2
feedgen
pyyaml
